refactoring and doc
