+++ mindsdb/api/mysql/mysql_proxy/datahub/datanodes/integration_datanode.py
@@ -11,6 +11,7 @@ from mindsdb_sql.parser.ast import Insert, Identifier, CreateTable, TableColumn,
 from mindsdb.api.mysql.mysql_proxy.datahub.datanodes.datanode import DataNode
 from mindsdb.api.mysql.mysql_proxy.libs.constants.response_type import RESPONSE_TYPE
 from mindsdb.api.mysql.mysql_proxy.datahub.classes.tables_row import TablesRow
+import mindsdb.utilities.profiler as profiler
 
 
 class DBHandlerException(Exception):
@@ -129,8 +130,8 @@ class IntegrationDataNode(DataNode):
         if result.type == RESPONSE_TYPE.ERROR:
             raise Exception(result.error_message)
 
+    @profiler.profile()
     def query(self, query=None, native_query=None, session=None):
-
         try:
             if query is not None:
                 result = self.integration_handler.query(query)
