+++ mindsdb/utilities/context.py
@@ -2,11 +2,6 @@
 from contextvars import ContextVar
 from typing import Any
 from copy import deepcopy
-import time
-from datetime import datetime, timezone
-
-import mindsdb.utilities.hooks as hooks
-from mindsdb.utilities.config import Config
 
 
 class Context:
@@ -22,6 +17,7 @@ class Context:
             'company_id': None,
             'user_class': 0,
             'profiling': {
+                'level': 0,
                 'enabled': False,
                 'pointer': None,
                 'tree': None
@@ -52,54 +48,6 @@ class Context:
 
     def load(self, storage: dict) -> None:
         self._storage.set(storage)
-
-    def _get_current_node(self, profiling):
-        current_node = profiling['tree']
-        for child_index in profiling['pointer']:
-            current_node = current_node['children'][child_index]
-        return current_node
-
-    def profiling_start_node(self, tag):
-        profiling = self.profiling
-        new_node = {
-            'start_at': time.perf_counter(),
-            'stop_at': None,
-            'name': tag,
-            'children': []
-        }
-        if profiling['pointer'] is None:
-            profiling['pointer'] = []
-            profiling['tree'] = new_node
-            profiling['tree']['time_start_at'] = datetime.now(timezone.utc)
-        else:
-            current_node = self._get_current_node(profiling)
-            profiling['pointer'].append(len(current_node['children']))
-            current_node['children'].append(new_node)
-
-    def profiling_set_meta(self, **kwargs):
-        self.profiling.update(kwargs)
-
-    def profiling_stop_current_node(self):
-        profiling = self.profiling
-        current_node = self._get_current_node(profiling)
-        current_node['stop_at'] = time.perf_counter()
-        if len(profiling['pointer']) > 0:
-            profiling['pointer'] = profiling['pointer'][:-1]
-        else:
-            self._send_profiling_results()
-            profiling['pointer'] = None
-
-    def _send_profiling_results(self):
-        self.profiling['company_id'] = self.company_id
-        self.profiling['hostname'] = Config().get('aws_meta_data', {}).get('public-hostname', '?')
-        self.profiling['instance_id'] = Config().get('aws_meta_data', {}).get('instance-id', '?')
-        hooks.send_profiling_results(self.profiling)
-
-    def enable_profiling(self):
-        self.profiling['enabled'] = True
-
-    def disable_profiling(self):
-        self.profiling['enabled'] = False
 
 
 _context_var = ContextVar('mindsdb.context')
