+++ mindsdb/utilities/profiler/__init__.py
@@ -1,0 +1,19 @@
+from mindsdb.utilities.profiler.profiler import (
+    start,
+    stop,
+    Context,
+    profile,
+    enable,
+    disable,
+    set_meta
+)
+
+__all__ = [
+    'start',
+    'stop',
+    'Context',
+    'profile',
+    'enable',
+    'disable',
+    'set_meta'
+]
